Job Scheduler UI (#636)
This change adds a new icon to schedule experiments to be run based on a cron pattern. Additionally, there is a new modal which connects to the endpoints to schedule experiments.  We also introduced a NEW dashboard that is oriented towards time series data for running the experiments over time.


---------

Signed-off-by: Anthony Leong <aj.leong623@gmail.com>
Signed-off-by: Eric Pugh <epugh@opensourceconnections.com>
Co-authored-by: Eric Pugh <epugh@opensourceconnections.com>